Refactor the current file structure
