+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:seg_coursework_app/pages/admin_home.dart';
+import 'package:seg_coursework_app/pages/admin_interface.dart';
 
 void main() {
   runApp(const AdminInterface());
